fix: 시간표 액티비티 예외 케이스 발생 시 온보딩 떨구기 보장
